pyproject.toml
[project]
name = "agentil-agent"
version = "0.3.0"
description = "Containerized agentic AI assistants"
readme = "README.md"
requires-python = ">=3.11,<3.12"
license = {text = "MIT"}
authors = [
    {name = "Hugo Morau", email = "morauhugo@gmail.com"}
]
keywords = ["agentil", "voice", "tts", "stt", "ai", "assistant", "websocket", "server"]

# Core dependencies
dependencies = [
    # OpenCode communication
    "httpx>=0.27.0",
    "httpx-sse>=0.4.0",
    
    # Configuration
    "pydantic>=2.5.0",
    "tomli>=2.0.0",
    "tomli-w>=1.0.0",
]

[project.optional-dependencies]
tts = [
    "numpy>=1.26.0",
    "melotts",
    "unidic==1.1.0",
    "soxr==0.5.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.0",
    "pydub>=0.25.0",
]
stt = [
    "numpy>=1.26.0",
    "openai-whisper>=20240930",
]

# Application bundles
server = [
    "agentil-agent[tts,stt]",
    
    # WebSocket Server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
]
client = [
    "rich>=13.7.0",
    "click>=8.1.0",
]

# Convenience
all = [
    "agentil-agent[voice]",
    "agentil-agent[websocket]"
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[project.scripts]
agentil-server = "agentil_agent.applications.server.main:main"
agentil-client = "agentil_agent.applications.cli_client.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentil_agent"]

[tool.uv.sources]
melotts = { git = "https://github.com/myshell-ai/MeloTTS.git" }

# PyTorch CUDA 12.8 - install manually via pip:
# pip install torch torchaudio --index-url https://download.pytorch.org/whl/nightly/cu128
# [[tool.uv.index]]
# name = "pytorch-cu128"
# url = "https://download.pytorch.org/whl/nightly/cu128"
# explicit = true

# [tool.uv.sources.torch]
# index = "pytorch-cu128"

# [tool.uv.sources.torchaudio]
# index = "pytorch-cu128"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]  # Line too long - handled by formatter

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
